requirements.txt
requests>=2.31.0
Pillow>=10.0.0
ollama>=0.1.0
